uncompressed: use explicit cast to silence clang tidy false alarm

+++ libheif/codecs/uncompressed_box.cc
@@ -486,7 +486,7 @@ Error Box_icef::write(StreamWriter& writer) const
   if (can_use_implied_offsets) {
     unit_offset_code = 0;
   }
-  uint8_t code_bits = (unit_offset_code << 5) | (unit_size_code << 2);
+  uint8_t code_bits = (uint8_t)((unit_offset_code << 5) | (unit_size_code << 2));
   writer.write8(code_bits);
   writer.write32((uint32_t)m_unit_infos.size());
   for (CompressedUnitInfo unit_info: m_unit_infos) {
